Improve error scenarios
Previously, if the message processor lambda function returned that an
error occurred while processing a message, the app would change the
message visibility to a hardcoded timeout of 5 seconds. However, SQS
already has a mechanism to deal with error scenarios: the queue's
visibility timeout setting.

This commit changes the app to no longer change message visibility for
messages that encountered an error. This allows clients to use the queue
visibility timeout to control how soon a message causing an error is
retried.

This commit also adds handling for if the message processor lambda
encounters an unhandled error (rather than returning a well-formed
output indicating only certain messages encountered errors).